Fixed some errors

+++ src/main/java/com/code/market/persistence/entity/Cliente.java
@@ -2,6 +2,7 @@ package com.code.market.persistence.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Entity
@@ -12,7 +13,7 @@ public class Cliente {
     private String id;
     private String nombre;
     private String apellidos;
-    private Integer celular;
+    private Long celular;
     private String direccion;
     @Column(name = "correo_electronico")
     private String correoElectronico;
@@ -44,11 +45,11 @@ public class Cliente {
         this.apellidos = apellidos;
     }
 
-    public Integer getCelular() {
+    public Long getCelular() {
         return celular;
     }
 
-    public void setCelular(Integer celular) {
+    public void setCelular(Long celular) {
         this.celular = celular;
     }
